Change qHash return value to size_t
Else on Windows there's a warning:
conversion from 'size_t' to 'uint', possible loss of data

+++ src/plugins/qmltooling/qmldbg_nativedebugger/qqmlnativedebugservice.cpp
@@ -83,7 +83,7 @@ public:
     int hitCount;
 };
 
-inline uint qHash(const BreakPoint &b, uint seed = 0) Q_DECL_NOTHROW
+inline size_t qHash(const BreakPoint &b, uint seed = 0) Q_DECL_NOTHROW
 {
     return qHash(b.fileName, seed) ^ b.lineNumber;
 }
